Widen RawCommit.commitTime to Long; restore Windows LTO.thin

+++ modules/resolution/src/main/scalajvm/version/resolution/jvm/JvmGitRepository.scala
@@ -185,7 +185,7 @@ final class JvmGitRepository private (repo: Repository) extends GitRepository:
       id = CommitSha(c.getId.name),
       message = c.getFullMessage.nn,
       parentIds = parentIds,
-      commitTime = c.getCommitTime
+      commitTime = c.getCommitTime.toLong
     )
 end JvmGitRepository
